requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.2
PyPDF2==1.26.0
pypng==0.0.21
PyQRCode==1.2.1
python-barcode==0.13.1
python-bidi==0.4.2
pytz==2021.1
tablib==3.0.0
typing-extensions==3.10.0.2
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
pydrive==1.3.1
python-dateutil==2.8.2

